Move tests into matching namespaces
